Admin and User Dashboard
Created adminMenu and userMenu that will be shown on dashboard page. In admin menu created links to create-category, create-product and users page(to display all users)
Created those pages and created routes for them and imported pages in app.js. In userMenu created orders and profile pages, which will be used to change profile info and track orders